rename lib to util
