--- lib/files.js
@@ -1,12 +1,0 @@
-const path = require('path');
-const fs = require('fs');
-
-module.exports = {
-    getCurrentDirectoryBase: () => {
-        return path.basename(process.cwd());
-    },
-
-    directoryExists: (filePath) => {
-        return fs.existsSync(filePath);
-    }
-};
